Solution to Find Maximum of Three Numbers

+++ Find Maximum of Three Numbers@@6788dd63490637cfc44e1b94/code.c
@@ -1,17 +1,17 @@
 #include <stdio.h>
 int main() {
-    int a , b , c;
-    scanf("%d",&a);
-    scanf("%d",&b);
-    scanf("%d",&c);
+    float a , b , c;
+    scanf("%f",&a);
+    scanf("%f",&b);
+    scanf("%f",&c);
     if(a>b && a>c) {
-        printf("%d",a);
+        printf("%f",a);
     }
     if(b>c && b>a) {
-        printf("%d",b);
+        printf("%f",b);
     }
     else {
-        printf("%d",c);
+        printf("%f",c);
     }
     return 0;
 }
